+++ lib/rspec/core/example_group.rb
@@ -33,19 +33,17 @@ module RSpec
         # @private
         def self.delegate_to_metadata(*names)
           names.each do |name|
-            define_method name do
-              metadata[:example_group][name]
-            end
+            define_method(name) { metadata.fetch(name) }
           end
         end
 
         # @return [String] the current example group description
         def description
-          description = metadata[:example_group][:description]
+          description = metadata[:description]
           RSpec.configuration.format_docstrings_block.call(description)
         end
 
-        delegate_to_metadata :described_class, :file_path
+        delegate_to_metadata :described_class, :file_path, :location
         alias_method :display_name, :description
         # @private
         alias_method :describes, :described_class
@@ -87,6 +85,10 @@ module RSpec
             options[:pending] = true
             reason = RSpec::Core::Pending::NO_REASON_GIVEN
           end
+
+          # Assign :caller so that the callback's source_location isn't used
+          # as the example location.
+          options[:caller] ||= Metadata.backtrace_from(block)
 
           # This will fail if no block is provided, which is effectively the
           # same as failing the example so it will be marked correctly as
@@ -179,7 +181,10 @@ module RSpec
         #   @see SharedExampleGroup
         def self.define_nested_shared_group_method(new_name, report_label="it should behave like")
           define_method(new_name) do |name, *args, &customization_block|
-            group = example_group("#{report_label} #{name}") do
+            # Pass :caller so the :location metadata is set properly...
+            # otherwise, it'll be set to the next line because that's
+            # the block's source_location.
+            group = example_group("#{report_label} #{name}", :caller => caller) do
               find_and_eval_shared("examples", name, *args, &customization_block)
             end
             group.metadata[:shared_group_name] = name
@@ -283,8 +288,6 @@ module RSpec
       # @see DSL#describe
       def self.example_group(*args, &example_group_block)
         args << {} unless args.last.is_a?(Hash)
-        args.last.update(:example_group_block => example_group_block)
-
         child = subclass(self, args, &example_group_block)
         children << child
         child
@@ -319,7 +322,7 @@ module RSpec
       # @private
       def self.subclass(parent, args, &example_group_block)
         subclass = Class.new(parent)
-        subclass.set_it_up(*args)
+        subclass.set_it_up(*args, &example_group_block)
         ExampleGroups.assign_const(subclass)
         subclass.module_eval(&example_group_block) if example_group_block
 
@@ -362,7 +365,7 @@ module RSpec
       end
 
       # @private
-      def self.set_it_up(*args)
+      def self.set_it_up(*args, &example_group_block)
         # Ruby 1.9 has a bug that can lead to infinite recursion and a
         # SystemStackError if you include a module in a superclass after
         # including it in a subclass: https://gist.github.com/845896
@@ -373,9 +376,12 @@ module RSpec
         ensure_example_groups_are_configured
 
         symbol_description = args.shift if args.first.is_a?(Symbol)
-        args << Metadata.build_hash_from(args)
+        user_metadata = Metadata.build_hash_from(args)
         args.unshift(symbol_description) if symbol_description
-        @metadata = RSpec::Core::Metadata.new(superclass_metadata).process(*args)
+        @metadata = Metadata::ExampleGroupHash.create(
+          superclass_metadata || {}, user_metadata, *args, &example_group_block
+        )
+
         hooks.register_globals(self, RSpec.configuration.hooks)
         world.configure_group(self)
       end
@@ -452,7 +458,7 @@ module RSpec
           warn <<-WARNING.gsub(/^ +\|/, '')
             |WARNING: Ignoring unknown ordering specified using `:order => #{order.inspect}` metadata.
             |         Falling back to configured global ordering.
-            |         Unrecognized ordering specified at: #{metadata[:example_group][:location]}
+            |         Unrecognized ordering specified at: #{location}
           WARNING
 
           registry.fetch(:global)
@@ -500,7 +506,7 @@ module RSpec
 
       # @private
       def self.declaration_line_numbers
-        @declaration_line_numbers ||= [metadata[:example_group][:line_number]] +
+        @declaration_line_numbers ||= [metadata[:line_number]] +
           examples.collect {|e| e.metadata[:line_number]} +
           children.inject([]) {|l,c| l + c.declaration_line_numbers}
       end