+++ spec/rspec/core/metadata_spec.rb
@@ -28,10 +28,17 @@ module RSpec
 
       context "when created" do
         Metadata::RESERVED_KEYS.each do |key|
-          it "prohibits :#{key} as a hash key" do
-            expect do
+          it "prohibits :#{key} as a hash key for an example group" do
+            expect {
               RSpec.describe("group", key => {})
-            end.to raise_error(/:#{key} is not allowed/)
+            }.to raise_error(/:#{key} is not allowed/)
+          end
+
+          it "prohibits :#{key} as a hash key for an example" do
+            group = RSpec.describe("group")
+            expect {
+              group.example("example", key => {})
+            }.to raise_error(/:#{key} is not allowed/)
           end
         end
 
@@ -42,7 +49,7 @@ module RSpec
             m = metadata
           end
 
-          expect(m[:example_group][:location]).to eq("example_file:42")
+          expect(m[:location]).to eq("example_file:42")
         end
       end
 
@@ -101,7 +108,7 @@ module RSpec
           expect(metadata_for("desc", :arbitrary => :options)).to have_value(:options).for(:arbitrary)
         end
 
-        it "points :example_group to the same hash object" do
+        it "points :example_group to the same hash object as other examples in the same group" do
           a = b = nil
 
           RSpec.describe "group" do
@@ -112,83 +119,112 @@ module RSpec
           a[:description] = "new description"
           expect(b[:description]).to eq("new description")
         end
+
+        it 'does not include example-group specific keys' do
+          metadata = nil
+
+          RSpec.describe "group" do
+            context "nested" do
+              metadata = example("foo").metadata
+            end
+          end
+
+          expect(metadata.keys).not_to include(:parent_example_group)
+        end
       end
 
-      [:described_class, :describes].each do |key|
-        describe key do
-          extract_key_from = lambda do |group|
-            group.metadata[:example_group][key]
+      describe ":block" do
+        context "for example group metadata" do
+          it "contains the example group block" do
+            block = Proc.new { }
+            group = RSpec.describe("group", &block)
+            expect(group.metadata[:block]).to equal(block)
+          end
+        end
+
+        context "for example metadata" do
+          it "contains the example block" do
+            block = Proc.new { }
+            group = RSpec.describe("group")
+            example = group.example("example", &block)
+            expect(example.metadata[:block]).to equal(block)
+          end
+        end
+      end
+
+      describe ":described_class" do
+        value_from = lambda do |group|
+          group.metadata[:described_class]
+        end
+
+        context "in an outer group" do
+          define_method :value_for do |arg|
+            value_from[RSpec.describe(arg)]
           end
 
-          context "in an outer group" do
-            define_method :value_for do |arg|
-              extract_key_from[RSpec.describe(arg)]
-            end
-
-            context "with a String" do
-              it "returns nil" do
-                expect(value_for "group").to be_nil
-              end
-            end
-
-            context "with a Symbol" do
-              it "returns nil" do
-                expect(value_for :group).to be_nil
-              end
-            end
-
-            context "with a class" do
-              it "returns the class" do
-                expect(value_for String).to be(String)
-              end
+          context "with a String" do
+            it "returns nil" do
+              expect(value_for "group").to be_nil
             end
           end
 
-          context "in a nested group" do
-            it "inherits the parent group's described class" do
-              value = nil
+          context "with a Symbol" do
+            it "returns nil" do
+              expect(value_for :group).to be_nil
+            end
+          end
 
-              RSpec.describe(Hash) do
+          context "with a class" do
+            it "returns the class" do
+              expect(value_for String).to be(String)
+            end
+          end
+        end
+
+        context "in a nested group" do
+          it "inherits the parent group's described class" do
+            value = nil
+
+            RSpec.describe(Hash) do
+              describe "sub context" do
+                value = value_from[self]
+              end
+            end
+
+            expect(value).to be(Hash)
+          end
+
+          it "sets the described class when passing a class" do
+            value = nil
+
+            RSpec.describe(String) do
+              describe Array do
+                value = value_from[self]
+              end
+            end
+
+            expect(value).to be(Array)
+          end
+
+          it "can override a parent group's described class using metdata" do
+            parent_value = child_value = grandchild_value = nil
+
+            RSpec.describe(String) do
+              parent_value = value_from[self]
+
+              describe "sub context" do
+                metadata[:described_class] = Hash
+                child_value = value_from[self]
+
                 describe "sub context" do
-                  value = extract_key_from[self]
+                  grandchild_value = value_from[self]
                 end
               end
-
-              expect(value).to be(Hash)
             end
 
-            it "sets the described class when passing a class" do
-              value = nil
-
-              RSpec.describe(String) do
-                describe Array do
-                  value = extract_key_from[self]
-                end
-              end
-
-              expect(value).to be(Array)
-            end
-
-            it "can override a parent group's described class using metdata" do
-              parent_value = child_value = grandchild_value = nil
-
-              RSpec.describe(String) do
-                parent_value = extract_key_from[self]
-
-                describe "sub context" do
-                  metadata[:example_group][key] = Hash
-                  child_value = extract_key_from[self]
-
-                  describe "sub context" do
-                    grandchild_value = extract_key_from[self]
-                  end
-                end
-              end
-
-              expect(grandchild_value).to be(Hash)
-              expect(child_value).to be(Hash)
-              expect(parent_value).to be(String)
-            end
+            expect(grandchild_value).to be(Hash)
+            expect(child_value).to be(Hash)
+            expect(parent_value).to be(String)
           end
         end
       end
@@ -211,7 +247,7 @@ module RSpec
             value = nil
 
             RSpec.describe(*args) do
-              value = metadata[:example_group][:description]
+              value = metadata[:description]
             end
 
             value
@@ -236,7 +272,7 @@ module RSpec
           end
 
           context "with empty args" do
-            it "returns empty string for [:example_group][:description]" do
+            it "returns empty string for [:description]" do
               expect(group_value_for()).to eq("")
             end
           end
@@ -261,7 +297,7 @@ module RSpec
 
           RSpec.describe "parent" do
             describe "child" do
-              group_value = metadata[:example_group][:full_description]
+              group_value = metadata[:full_description]
               example_value = example("example").metadata[:full_description]
             end
           end
@@ -274,11 +310,11 @@ module RSpec
           grandparent_value = parent_value = child_value = example_value = nil
 
           RSpec.describe "grandparent" do
-            grandparent_value = metadata[:example_group][:full_description]
+            grandparent_value = metadata[:full_description]
             describe "parent" do
-              parent_value = metadata[:example_group][:full_description]
+              parent_value = metadata[:full_description]
               describe "child" do
-                child_value = metadata[:example_group][:full_description]
+                child_value = metadata[:full_description]
                 example_value = example("example").metadata[:full_description]
               end
             end
@@ -296,7 +332,7 @@ module RSpec
               value = nil
 
               RSpec.describe Array, "#{char}method" do
-                value = metadata[:example_group][:full_description]
+                value = metadata[:full_description]
               end
 
               expect(value).to eq("Array#{char}method")
@@ -309,7 +345,7 @@ module RSpec
 
               RSpec.describe Object do
                 describe "#{char}method" do
-                  value = metadata[:example_group][:full_description]
+                  value = metadata[:full_description]
                 end
               end
 
@@ -324,7 +360,7 @@ module RSpec
               RSpec.describe(Array) do
                 context "with 2 items" do
                   describe "#{char}method" do
-                    value = metadata[:example_group][:full_description]
+                    value = metadata[:full_description]
                   end
                 end
               end
@@ -340,7 +376,7 @@ module RSpec
           value = nil
 
           RSpec.describe(:caller => ["./lib/rspec/core/foo.rb", "#{__FILE__}:#{__LINE__}"]) do
-            value = metadata[:example_group][:file_path]
+            value = metadata[:file_path]
           end
 
           expect(value).to eq(relative_path(__FILE__))
@@ -353,7 +389,7 @@ module RSpec
 
           @describe_line = __LINE__ + 1
           RSpec.describe(*args) do
-            value = metadata[:example_group][:line_number]
+            value = metadata[:line_number]
           end
 
           value
@@ -381,7 +417,130 @@ module RSpec
             describe { child = metadata }
           end
 
-          expect(child[:example_group][:example_group]).to eq(parent[:example_group])
+          expect(child[:parent_example_group]).to eq(parent)
+        end
+      end
+
+      describe "backwards compatibility" do
+        before { allow_deprecation }
+
+        describe ":example_group" do
+          it 'issues a deprecation warning when the `:example_group` key is accessed' do
+            expect_deprecation_with_call_site(__FILE__, __LINE__ + 2, /:example_group/)
+            RSpec.describe(Object, "group") do
+              metadata[:example_group]
+            end
+          end
+
+          it 'can still access the example group attributes via [:example_group]' do
+            meta = nil
+            RSpec.describe(Object, "group") { meta = metadata }
+
+            expect(meta[:example_group][:line_number]).to eq(__LINE__ - 2)
+            expect(meta[:example_group][:description]).to eq("Object group")
+          end
+
+          it 'can access the parent example group attributes via [:example_group][:example_group]' do
+            parent = child = nil
+            parent_line = __LINE__ + 1
+            RSpec.describe(Object, "group", :foo => 3) do
+              parent = metadata
+              describe("nested") { child = metadata }
+            end
+
+            expect(child[:example_group][:example_group]).to include(
+              :foo => 3,
+              :description => "Object group",
+              :line_number => parent_line
+            )
+          end
+
+          it 'can mutate attributes when accessing them via [:example_group]' do
+            meta = nil
+
+            RSpec.describe(String) do
+              describe "sub context" do
+                meta = metadata
+              end
+            end
+
+            expect {
+              meta[:example_group][:described_class] = Hash
+            }.to change { meta[:described_class] }.from(String).to(Hash)
+          end
+
+          it 'can still be filtered via a nested key under [:example_group] as before' do
+            meta = nil
+
+            line = __LINE__ + 1
+            RSpec.describe("group") { meta = metadata }
+
+            applies = MetadataFilter.any_apply?(
+              { :example_group => { :line_number => line } },
+              meta
+            )
+
+            expect(applies).to be true
+          end
+        end
+
+        describe ":example_group_block" do
+          it 'returns the block' do
+            meta = nil
+
+            RSpec.describe "group" do
+              meta = metadata
+            end
+
+            expect(meta[:example_group_block]).to be_a(Proc).and eq(meta[:block])
+          end
+
+          it 'issues a deprecation warning' do
+            expect_deprecation_with_call_site(__FILE__, __LINE__ + 2, /:example_group_block/)
+            RSpec.describe "group" do
+              metadata[:example_group_block]
+            end
+          end
+        end
+
+        describe ":describes" do
+          context "on an example group metadata hash" do
+            it 'returns the described_class' do
+              meta = nil
+
+              RSpec.describe Hash do
+                meta = metadata
+              end
+
+              expect(meta[:describes]).to be(Hash).and eq(meta[:described_class])
+            end
+
+            it 'issues a deprecation warning' do
+              expect_deprecation_with_call_site(__FILE__, __LINE__ + 2, /:describes/)
+              RSpec.describe "group" do
+                metadata[:describes]
+              end
+            end
+          end
+
+          context "an an example metadata hash" do
+            it 'returns the described_class' do
+              meta = nil
+
+              RSpec.describe Hash do
+                meta = example("ex").metadata
+              end
+
+              expect(meta[:describes]).to be(Hash).and eq(meta[:described_class])
+            end
+
+            it 'issues a deprecation warning' do
+              expect_deprecation_with_call_site(__FILE__, __LINE__ + 2, /:describes/)
+              RSpec.describe "group" do
+                example("ex").metadata[:describes]
+              end
+            end
+          end
         end
       end
     end